+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fix `ImageSource`, `VideoSource` and `CanvasSource` leaking a GPU texture on every image update and on removal, and a resized texture losing its wrap and filter settings ([#8094](https://github.com/maplibre/maplibre-gl-js/pull/8094)) (by [@mondsichtung](https://github.com/mondsichtung))
 - Fix `map.queryRenderedFeatures()` sometimes causing "Out of bounds" error due to race condition while loading tile data ([#8064](https://github.com/maplibre/maplibre-gl-js/issues/8064)) (by [@smvjohansenbouvet](https://github.com/smvjohansenbouvet))
 - Fix zooming the globe with the scroll wheel or a two-finger pinch drifting away from the pointer while the globe is small on screen, instead of keeping the location under the pointer as it does when zoomed in ([#8095](https://github.com/maplibre/maplibre-gl-js/pull/8095)) (by [@mondsichtung](https://github.com/mondsichtung))
+- Fix `dragend` and `moveend` reporting a different center and zoom when terrain DEM elevation arrives during a drag ([#7989](https://github.com/maplibre/maplibre-gl-js/issues/7989)) (by [@xvodddwannaG](https://github.com/xvodddwannaG))
 - _...Add new stuff here..._
 
 ## 6.2.0